requirements.txt
numpy==1.26.2
opencv-contrib-python==4.8.1.78
Pillow==10.1.0
PyYAML==6.0.1
tqdm==4.66.1
